html/semantics/document-metadata/the-link-element/link-dcl-event.html

<!DOCTYPE html>
<link rel="author" title="Dom Farolino" href="mailto:dom@chromium.org">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#the-link-element">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
var saw_link_onload = false;
var t = async_test("Check if the stylesheet's load event blocks the " +
                   "DOMContentLoaded event when it precedes the pending " +
                   "parsing-blocking script");
window.addEventListener('DOMContentLoaded', t.step_func_done(function() {
  assert_true(saw_link_onload, "The DOMContentLoaded event should not fire " +
                               "before a link element that contributes a " +
                               "script-blocking style sheet's load event is " +
                               "fired when it precedes the pending " +
                               "parsing-blocking script");
}));
</script>
<link href="style.css?pipe=trickle(d3)" rel="stylesheet" id="style_test"
      onload="t.step(function() { saw_link_onload = true; });"
      onerror="t.step(function() { assert_unreached('Sheet should load OK'); });"></link>
<script>var nop = 0;</script>
</head>
</html>
